requirements.txt
asgiref==3.2.10
Django==3.1.2
djangorestframework==3.12.1
psycopg2==2.8.6
Pygments==2.7.2
pytz==2020.1
sqlparse==0.4.1
